1.修改订单支付的默认支付方式为余额支付. 2.下单界面使用支付宝支付流程完成. 3.订单完成后的详情页面,充电耗时设定为服务器传回来的power_time 4.充电过程中增加实时和服务器的长连接.添加第三方的依赖库

+++ app/src/main/java/com/optimumnano/quickcharge/bean/RechargeGunBean.java
@@ -11,5 +11,5 @@ public class RechargeGunBean {
     public String pile_type;
     public int power;
     public int elec_current;
-    public int price;
+    public Double price;
 }
